fix: dogecoin_pubkey.compressed is uint8_t not int; add ECC context guards to key tests
dogecoin_bool is typedef'd to uint8_t in dogecoin.h.  Declaring the
compressed field as int in tier3_preamble.h made cffi reject ffi.new()
with a size mismatch (cdef says 4, C compiler says 1).  Change the field
to uint8_t so the cdef layout matches the ABI.

Also add setup_module/teardown_module ECC context guards to key_test.py
(same pattern as hdnode_test.py) so Linux debug assertions do not abort
when the test file runs after address_test.py tears down the context.

+++ codegen/tier3_preamble.h
@@ -44,7 +44,7 @@ typedef struct {
 } dogecoin_key;
 
 typedef struct {
-    int compressed;          /* dogecoin_bool */
+    uint8_t compressed;      /* dogecoin_bool = uint8_t */
     uint8_t pubkey[65];
 } dogecoin_pubkey;
